requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.2.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dm-tree==0.1.8
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lightning-utilities==0.10.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
namex==0.0.7
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pexpect==4.8.0
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.23.4
ptyprocess==0.7.0
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==7.4.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytorch-lightning==2.1.2
pytz==2022.1
PyYAML==5.4.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sos==4.4
ssh-import-id==5.11
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
Twisted==22.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.24.0.post1
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
xkit==0.0.0
yarl==1.9.3
zipp==1.0.0
zope.interface==5.4.0
